% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HubBSP2-package.R
\docType{package}
\name{HubBSP2-package}
\alias{HubBSP2-package}
\alias{HubBSP2}
\title{Provide access to BrainSEQ Phase II processed data objects.}
\description{
This is an example ExperimentHub package made live during an LIBD
rstats club session. It provides access to the BrainSEQ Phase II
RangedSummarizedExperiment object at the gene-level. See
http://eqtl.brainseq.org/phase2/ for more information.
}
\author{
Leonardo Collado-Torres
}
